Begin moving common bits to a trait
