+++ readme.txt
@@ -3,6 +3,7 @@ Contributors: ethitter, automattic
 Tags: cron, bulk edit, bulk actions
 Requires at least: 4.8.1
 Tested up to: 4.9
+Required PHP: 5.4
 Stable tag: 1.0
 License: GPLv2 or later
 License URI: http://www.gnu.org/licenses/gpl-2.0.html
